Heterogenous action list view
Add inline headings to the action list view.
The listview delegate for the action list is now in its own QML component
so that multiple delegates can be used for header and actions based on a
flag in the data model checked in a Loader component.
Multiple delegates will be added for the different action orderings.