Enable loading of downloaded task data into the QML list view
Add downloading of the resources json data to fetch context and project
names by uuid.
Building of the list model data in the required display order from the
downloaded and parsed json data is performed in one place in C++ for
display in QML and later saving to file.
Only order by box is implemented currently.