Re-order and reload model data when different ordering selected
When user selects a different "Order by" radio button, the QML calls into C++
to re-order the data model, then reloads it into the QML ListModel.
This is done at startup as the default ordering type is set when intialising,
so the dummy list model data set is no longer shown nor needed.